refactor: consolidating separate packages
consolidating separate packages into main package for improved user-friendliness
