+++ scrape.go
@@ -1,6 +1,7 @@
-package scrape
+package siamauth
 
 import (
+	"fmt"
 	"io"
 	"log"
 	"regexp"
@@ -8,15 +9,11 @@ import (
 	"strings"
 
 	"github.com/PuerkitoBio/goquery"
-	"github.com/dnabil/siamauth/siamerr"
-	models "github.com/dnabil/siamauth/siammodel"
-	"github.com/dnabil/siamauth/util"
 )
 
-
-
-func ScrapeAddCourse(r io.Reader) ([]models.AddCourse, error) {
-	var courses []models.AddCourse
+// page: addcourse
+func ScrapeAddCourse(r io.Reader) ([]AddCourse, error) {
+	var courses []AddCourse
 
 	doc, err := goquery.NewDocumentFromReader(r)
 	if err != nil {
@@ -27,7 +24,7 @@ func ScrapeAddCourse(r io.Reader) ([]models.AddCourse, error) {
 	tableHeader := doc.Find("tr.textWhite")
 	thLength := tableHeader.Length()
 	if thLength <= 0 {
-		return courses, siamerr.ErrNoElement
+		return courses, ErrNoElement
 	}
 
 	// iterate to find the right table header
@@ -44,7 +41,7 @@ func ScrapeAddCourse(r io.Reader) ([]models.AddCourse, error) {
 		}
 	}
 	if !found {
-		return courses, siamerr.ErrNoElement
+		return courses, ErrNoElement
 	}
 
 	tBody := tableHeader.Parent()
@@ -52,21 +49,21 @@ func ScrapeAddCourse(r io.Reader) ([]models.AddCourse, error) {
 	// == scrape ==
 	// -scraping masa krs, ex: GANJIL 2023/2024, GENAP 2023/2024
 	span := doc.Find("span.section")
-	if !strings.EqualFold("Jadwal Mata Kuliah Ditawarkan", util.TrimSpace(span.Text())){
-		return courses, siamerr.ErrNoElement
+	if !strings.EqualFold("Jadwal Mata Kuliah Ditawarkan", trimSpace(span.Text())){
+		return courses, ErrNoElement
 	}
 
 	MasaKrsText := span.Parent().Text() // [string] : [MASA KRS]
 	masaArr := strings.Split(MasaKrsText, ":")
 	if len(masaArr) < 2 {
-		return courses, siamerr.ErrNoElement
+		return courses, ErrNoElement
 	}
 
-	masaKrs := util.TrimSpace(masaArr[1])
+	masaKrs := trimSpace(masaArr[1])
 	// end of scraping masa krs
 
 	trs := tBody.Find("tr:nth-of-type(n+2)") // skip table header
-	courses = make([]models.AddCourse, 0, trs.Length())
+	courses = make([]AddCourse, 0, trs.Length())
 
 	trs.Each(func(i int, s *goquery.Selection) {
 		// the tds are:
@@ -78,7 +75,7 @@ func ScrapeAddCourse(r io.Reader) ([]models.AddCourse, error) {
 		// put it into array first, proccess it then append the data to courses.
 
 		for i := 0; i < len(tempArr); i++ {
-			tempArr[i] = util.TrimSpace(tds.Eq(i).Text())
+			tempArr[i] = trimSpace(tds.Eq(i).Text())
 		}
 
 		kelasProdi := regexp.MustCompile(`^(\S+)\s+(.*)`).FindStringSubmatch(tempArr[2])
@@ -97,7 +94,7 @@ func ScrapeAddCourse(r io.Reader) ([]models.AddCourse, error) {
 
 		sks, _ := strconv.Atoi(tempArr[7])
 
-		courses = append(courses, models.AddCourse{
+		courses = append(courses, AddCourse{
 			MasaKRS: masaKrs,
 			Hari:         tempArr[0],
 			Jam:          tempArr[1],
@@ -114,6 +111,66 @@ func ScrapeAddCourse(r io.Reader) ([]models.AddCourse, error) {
 		})
 	})
 	// == end of scraping ==
-
 	return courses, nil
+}
+
+// page: akademik (dashboard)
+func ScrapeDataUser(r io.Reader) (UserData, error) {
+	doc, err := goquery.NewDocumentFromReader(r)
+	if err != nil {
+		return UserData{}, err
+	}
+
+	bioElement := doc.Find("div[class=\"bio-info\"]")
+	if bioElement.Length() == 0 {
+		return UserData{}, ErrNoElement
+	}
+
+	result := make([]string, 8)
+	divs := bioElement.Children()
+
+	divs.Each(func(i int, s *goquery.Selection) {
+		each := trimSpace(s.Text())
+		if each != "PDDIKTI KEMDIKBUDDetail" {
+			result[i] = each
+		}
+	})
+
+	userData := UserData{}
+
+	userData.NIM = trimSpace(result[0])
+	userData.Nama = trimSpace(result[1])
+	// result2 = Jenjang/Fakultas--/--
+	jenjangFakultas := strings.Split(result[2][16:], "/")
+	userData.Jenjang = jenjangFakultas[0]
+	userData.Fakultas = jenjangFakultas[1]
+	userData.Jurusan = trimSpace(result[3][7:])
+	userData.ProgramStudi = trimSpace(result[4][13:])
+	userData.Seleksi = trimSpace(result[5][7:])
+	fmt.Println(userData.Seleksi)
+	userData.NomorUjian = trimSpace(result[6][11:])
+	userData.FotoProfil = fmt.Sprintf("https://siakad.ub.ac.id/dirfoto/foto/foto_20%s/%s.jpg", userData.NIM[0:2], userData.NIM)
+
+	return userData, nil
+}
+
+// page: login 
+// 
+// will return error login message as string and an error
+//
+// error may be caused by: html parse error, siam UI changes
+func ScrapeLoginError(r io.Reader) (string, error) {
+	doc, err := goquery.NewDocumentFromReader(r)
+	if err != nil {
+		return "", err
+	}
+
+	msgElement := doc.Find("small.error-code")
+	if msgElement.Length() == 0 {
+		return "", ErrNoElement
+	}
+
+	msgString := trimSpace(msgElement.Text())
+	
+	return msgString, nil
 }
