+++ siamauth.go
@@ -3,25 +3,24 @@ package siamauth
 import (
 	"bytes"
 
-	"github.com/dnabil/siamauth/scrape"
-	"github.com/dnabil/siamauth/siamerr"
-	models "github.com/dnabil/siamauth/siammodel"
 	"github.com/gocolly/colly"
 )
 
 var (
 	loginPath		string = "index.php"
 	logoutPath		string = "logout.php"
+	// addCoursePath	string = "addcourse.php"
 
 	siamUrl			string = "https://siam.ub.ac.id/"			//GET
 	loginUrl		string = siamUrl + loginPath				//POST
 	logoutUrl		string = siamUrl + logoutPath				//GET
+	// addCourseUrl	string = siamUrl + addCoursePath
 )
 
 type (
 	User struct {
 		c           *colly.Collector
-		Data     	models.UserData
+		Data     	UserData
 		LoginStatus bool
 	}
 )
@@ -37,19 +36,19 @@ func NewUser() *User {
 // no need to login first & defer logout.
 //
 // if you just need to get the data and bounce, use this ;)
-func (s *User) GetDataAndLogout(username, password string) (models.UserData, error) {
+func (s *User) GetDataAndLogout(username, password string) (UserData, error) {
 	errMsg, err := s.Login(username, password)
 	if err != nil {
-		return models.UserData{}, err
+		return UserData{}, err
 	}
 	if errMsg != "" {
-		return models.UserData{}, siamerr.ErrLoginFail
+		return UserData{}, ErrLoginFail
 	}
 	defer s.Logout()
 
 	err = s.GetData()
 	if err != nil {
-		return models.UserData{}, err
+		return UserData{}, err
 	}
 
 	return s.Data, nil
@@ -61,7 +60,7 @@ func (s *User) GetDataAndLogout(username, password string) (models.UserData, err
 // will return a login error message (from siam) and an error (already logged in/login error/siam ui changes/server down/etc)
 func (s *User) Login(us string, ps string) (string, error) {
 	if s.LoginStatus {
-		return "", siamerr.ErrLoggedIn
+		return "", ErrLoggedIn
 	}
 
 	var errOnResponse error
@@ -73,7 +72,7 @@ func (s *User) Login(us string, ps string) (string, error) {
 		// may visit this path if login failed
 		if r.FileName() == loginPath {
 			response := bytes.NewReader(r.Body)
-			loginErrMsg, errOnResponse = scrape.ScrapeLoginError(response)
+			loginErrMsg, errOnResponse = ScrapeLoginError(response)
 		}
 	})
 
@@ -94,7 +93,7 @@ func (s *User) Login(us string, ps string) (string, error) {
 
 	// login fail
 	if loginErrMsg != "" {
-		return loginErrMsg, siamerr.ErrLoginFail
+		return loginErrMsg, ErrLoginFail
 	}
 
 	return "", nil
@@ -103,11 +102,11 @@ func (s *User) Login(us string, ps string) (string, error) {
 // GetData will fill in user's Data or return an error
 func (s *User) GetData() error {
 	var onScrapeErr error
-	var data models.UserData
+	var data UserData
 
 	// scraping data mahasiswas
 	s.c.OnHTML("", func(h *colly.HTMLElement) {
-		data, onScrapeErr = scrape.ScrapeDataUser(bytes.NewReader(h.Response.Body))
+		data, onScrapeErr = ScrapeDataUser(bytes.NewReader(h.Response.Body))
 	})
 	err := s.c.Visit(siamUrl)
 	if err != nil { return err }
@@ -120,7 +119,7 @@ func (s *User) GetData() error {
 // Make sure to defer this method after login, so the phpsessionid won't be misused
 func (s *User) Logout() error {
 	if !s.LoginStatus {
-		return siamerr.ErrNotLoggedIn
+		return ErrNotLoggedIn
 	}
 	s.c.Visit(logoutUrl)
 	return nil
