+++ util.go
@@ -1,7 +1,7 @@
-package util
+package siamauth
 
 import "strings"
 
-func TrimSpace(s string) string {
+func trimSpace(s string) string {
 	return strings.Join(strings.Fields(s), " ")
 }
